requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-rq==2.10.2
pillow==11.3.0
python-dotenv==1.1.1
psycopg2-binary==2.9.9
django-redis==5.4.0
gunicorn==22.0.0
redis==5.0.8
pytest==8.3.2
pytest-django==4.9.0
rq==1.16.2
whitenoise==6.7.0